Refactor SemIRDiagnosticConverter out of check.cpp
